requirements.txt
fastapi
uvicorn
requests
pydantic
sqlalchemy
gunicorn
python-multipart
africastalking
python-dotenv